module_1/requirements.txt
flask>=2.3.0
pymongo>=4.0.0
dnspython>=2.0.0
werkzeug>=2.3.0
python-dotenv>=1.0.0